+++ src/components/details-view/details-view.test.js
@@ -1,7 +1,6 @@
 import {
   render,
   fireEvent,
-  act,
 } from '@testing-library/react';
 
 import { createMemoryHistory } from 'history';
@@ -255,56 +254,6 @@ describe('Given DetailsView', () => {
       fireEvent.click(closeIcon);
 
       expect(history.location.pathname).toBe('/eholdings');
-    });
-  });
-
-  describe('when window is resized', () => {
-    const mockScrollIntoView = jest.fn();
-
-    afterEach(() => {
-      mockScrollIntoView.mockClear();
-    });
-
-    it('should handle layout changes and scroll', () => {
-      window.HTMLElement.prototype.scrollIntoView = mockScrollIntoView;
-
-      const { getByTestId } = renderDetailsView({
-        renderList: mockRenderList,
-        sections: { showTestSection: true },
-        listSectionId: 'showTestSection',
-        resultsLength: 6,
-        listType: 'testListType',
-      });
-
-      act(() => {
-        global.innerWidth = 1200;
-        global.innerHeight = 800;
-        global.dispatchEvent(new Event('resize'));
-        fireEvent.scroll(getByTestId('scroll-container'), { y: 500 });
-      });
-
-      expect(mockRenderList).toHaveBeenCalledWith(true);
-      expect(mockScrollIntoView).toHaveBeenCalled();
-    });
-
-    it('should handle layout changes but not scroll', () => {
-      window.HTMLElement.prototype.scrollIntoView = mockScrollIntoView;
-
-      const { getByTestId } = renderDetailsView({
-        renderList: mockRenderList,
-        resultsLength: 6,
-        listType: 'testListType',
-      });
-
-      act(() => {
-        global.innerWidth = 1200;
-        global.innerHeight = 800;
-        global.dispatchEvent(new Event('resize'));
-        fireEvent.scroll(getByTestId('scroll-container'), { y: 500 });
-      });
-
-      expect(mockRenderList).toHaveBeenCalledWith(true);
-      expect(mockScrollIntoView).not.toHaveBeenCalled();
     });
   });
 });